add generic loss metric (#112)
* add generic loss metric

* mark variables as volatile in Loss metric
