+++ examples/mnist_with_visdom.py
@@ -15,7 +15,7 @@ except ImportError:
     raise RuntimeError("No visdom package is found. Please install it with command: \n pip install visdom")
 
 from ignite.engines import Events, create_supervised_trainer, create_supervised_evaluator
-from ignite.metrics import CategoricalAccuracy, NegativeLogLikelihood
+from ignite.metrics import CategoricalAccuracy, Loss
 
 
 class Net(nn.Module):
@@ -64,10 +64,10 @@ def run(train_batch_size, val_batch_size, epochs, lr, momentum, log_interval):
     if cuda:
         model = model.cuda()
     optimizer = SGD(model.parameters(), lr=lr, momentum=momentum)
-    trainer = create_supervised_trainer(model, optimizer, nn.NLLLoss(), cuda=cuda)
+    trainer = create_supervised_trainer(model, optimizer, F.nll_loss, cuda=cuda)
     evaluator = create_supervised_evaluator(model,
                                             metrics={'accuracy': CategoricalAccuracy(),
-                                                     'nll': NegativeLogLikelihood()},
+                                                     'nll': Loss(F.nll_loss)},
                                             cuda=cuda)
 
     train_loss_window = create_plot_window(vis, '#Iterations', 'Loss', 'Training Loss')
