+++ ignite/metrics/__init__.py
@@ -1,9 +1,9 @@
 from .binary_accuracy import BinaryAccuracy
 from .categorical_accuracy import CategoricalAccuracy
+from .loss import Loss
 from .mean_absolute_error import MeanAbsoluteError
 from .mean_pairwise_distance import MeanPairwiseDistance
 from .mean_squared_error import MeanSquaredError
 from .metric import Metric
 from .root_mean_squared_error import RootMeanSquaredError
 from .top_k_categorical_accuracy import TopKCategoricalAccuracy
-from .negative_log_likelihood import NegativeLogLikelihood
